// frontend/src/types/admin.types.ts
export interface AdminUser {
  id: number;
  full_name: string;
  email: string;
  mobile_number: string;
  role: UserRole;
  is_verified: boolean;
  is_active: boolean;
  is_suspended: boolean;
  created_at: string;
}